refactor: move SEO logic to seoHelpers module
